pyproject.toml
[project]
name = "bedrock"
version = "0.1.0"
description = "Bedrock guide for AWS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.15",
    "load-dotenv>=0.1.0",
    "numpy>=2.3.2",
]
